+++ assets/js/gift-wrap-blocks.asset.php
@@ -1,1 +1,1 @@
-<?php return array('dependencies' => array('react-jsx-runtime', 'wc-blocks-checkout', 'wc-settings', 'wp-data', 'wp-element', 'wp-plugins'), 'version' => 'accc0d8e2a2381864943');
+<?php return array('dependencies' => array('react-jsx-runtime', 'wc-blocks-checkout', 'wc-settings', 'wp-data', 'wp-element', 'wp-plugins'), 'version' => '24f71b9b1433e0f4cbb6');